+++ compiler/ml/typecore.ml
@@ -50,7 +50,8 @@ type error =
   | Undefined_method of type_expr * string * string list option
   | Private_type of type_expr
   | Private_label of Longident.t * type_expr
-  | Not_subtype of (type_expr * type_expr) list * (type_expr * type_expr) list
+  | Not_subtype of
+      Ctype.type_pairs * Ctype.type_pairs * Ctype.subtype_context option
   | Too_many_arguments of bool * type_expr
   | Abstract_wrong_label of Noloc.arg_label * type_expr
   | Scoping_let_module of string * type_expr
@@ -601,8 +602,8 @@ let extract_type_from_pat_variant_spread env lid expected_ty =
       raise (Error (lid.loc, env, Type_params_not_supported lid.txt));
     let ty = newgenty (Tconstr (path, [], ref Mnil)) in
     (try Ctype.subtype env ty expected_ty ()
-     with Ctype.Subtype (tr1, tr2) ->
-       raise (Error (lid.loc, env, Not_subtype (tr1, tr2))));
+     with Ctype.Subtype (tr1, tr2, ctx) ->
+       raise (Error (lid.loc, env, Not_subtype (tr1, tr2, ctx))));
     (path, decl, constructors, ty)
   | _ -> raise (Error (lid.loc, env, Not_a_variant_type lid.txt))
 
@@ -2933,9 +2934,9 @@ and type_expect_ ?type_clash_context ?in_function ?(recarg = Rejected) env sexp
            let force' = subtype env arg.exp_type ty' in
            force ();
            force' ()
-         with Subtype (tr1, tr2) ->
+         with Subtype (tr1, tr2, ctx) ->
            (* prerr_endline "coercion failed"; *)
-           raise (Error (loc, env, Not_subtype (tr1, tr2))));
+           raise (Error (loc, env, Not_subtype (tr1, tr2, ctx))));
       (arg, ty', cty')
     in
     rue
@@ -4329,8 +4330,8 @@ let report_error env loc ppf error =
     match valid_methods with
     | None -> ()
     | Some valid_methods -> spellcheck ppf me valid_methods)
-  | Not_subtype (tr1, tr2) ->
-    report_subtyping_error ppf env tr1 "is not a subtype of" tr2
+  | Not_subtype (tr1, tr2, ctx) ->
+    report_subtyping_error ppf env tr1 "is not a subtype of" tr2 ctx
   | Too_many_arguments (in_function, ty) ->
     if
       (* modified *)